feat: make language selector be a component
